refactor: extract logic to NoteGenerator.java that generate notes to excel
